Initial support for static library
