+++ include/spdlog/spdlog.h
@@ -152,4 +152,4 @@ void drop_all();
 }
 
 
-#include "details/spdlog_impl.h"
+#include "./details/spdlog_impl.h"
