+++ include/spdlog/tweakme.h
@@ -72,3 +72,9 @@
 // Note that upon creating a logger the registry is modified by spdlog..
 // #define SPDLOG_NO_REGISTRY_MUTEX
 ///////////////////////////////////////////////////////////////////////////////
+
+
+///////////////////////////////////////////////////////////////////////////////
+// Uncomment if you want to compile as a static library
+// #define SPDLOG_LIBRARY
+///////////////////////////////////////////////////////////////////////////////
