+++ tests/includes.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdio>
+#include <cstring>
 #include <fstream>
 #include <string>
 #include <ostream>
@@ -10,3 +11,5 @@
 #include "catch.hpp"
 #include "../include/spdlog/spdlog.h"
 #include "../include/spdlog/sinks/null_sink.h"
+#include "../include/spdlog/sinks/file_sinks.h"
+#include "../include/spdlog/sinks/ostream_sink.h"
